// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/getTemplateName1SourceType.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz
 * OpenAPI spec version: v1.0.0
 */

export type GetTemplateName1SourceType =
  (typeof GetTemplateName1SourceType)[keyof typeof GetTemplateName1SourceType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetTemplateName1SourceType = {
  LOCAL_DIR: "LOCAL_DIR",
  EMAIL: "EMAIL",
  S3: "S3",
  SFTP: "SFTP",
} as const;