// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/transferSourceCheckpointItemViewDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz
 * OpenAPI spec version: v1.0.0
 */
import type { Object } from "./object";

export interface TransferSourceCheckpointItemViewDTO {
  /** 去重记录主键。 */
  checkpointItemId?: string;
  /** 创建时间。 */
  createdAt?: string;
  /** 条目指纹。 */
  itemFingerprint?: string;
  /** 去重键。 */
  itemKey?: string;
  /** 条目元数据。(object) */
  itemMeta?: Object;
  /** 条目 MIME 类型。 */
  itemMimeType?: string;
  /** 条目名称。 */
  itemName?: string;
  /** 来源引用标识。 */
  itemRef?: string;
  /** 条目大小。 */
  itemSize?: number;
  /** 处理时间。 */
  processedAt?: string;
  /** 来源主键。 */
  sourceId?: string;
  /** 来源类型。 */
  sourceType?: string;
  /** 触发类型。 */
  triggerType?: string;
  /** 修改时间。 */
  updatedAt?: string;
}