// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/transferSourceCheckpointViewDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz
 * OpenAPI spec version: v1.0.0
 */
import type { Object } from "./object";

export interface TransferSourceCheckpointViewDTO {
  /** 检查点主键。 */
  checkpointId?: string;
  /** 检查点键。 */
  checkpointKey?: string;
  /** 检查点元数据。(object) */
  checkpointMeta?: Object;
  /** 检查点值。 */
  checkpointValue?: string;
  /** 创建时间。 */
  createdAt?: string;
  /** 来源主键。 */
  sourceId?: string;
  /** 来源类型。 */
  sourceType?: string;
  /** 修改时间。 */
  updatedAt?: string;
}